Update interface and enum naming convention, rename project overview enum
